feat: Add Customer Data Service with unified identity and consent management
- Introduced a new service for managing customer data, including profiles, consents, and risk-based authentication.
- Implemented API endpoints for profile management, consent handling, and risk provider integration.
- Added internal smoke tests for service validation during startup.
- Enhanced configuration management and error handling in the service.
- Created necessary data structures and JSON serialization methods for service operations.
- Updated build configurations and dependencies for the new service.

+++ customer-data/source/uim/sap/cdc/service.d
@@ -72,7 +72,11 @@ class CDCService {
       "email_verified",
       existing.isNull ? false : existing.get.emailVerified
     );
-    profile.preferences = readObject(body, "preferences", existing.isNull ? Json.emptyObject : existing.get.preferences);
+    profile.preferences = readObject(
+      body,
+      "preferences",
+      existing.isNull ? Json.emptyObject : existing.get.preferences
+    );
     profile.customAttributes = readObject(
       body,
       "custom_attributes",
@@ -324,7 +328,15 @@ class CDCService {
       }
 
       _store.upsertProfile(profile);
-      return deniedAuthPayload(tenantId, userId, ipAddress, "deny_invalid_credentials", riskLevel, riskScore, providerSignals);
+      return deniedAuthPayload(
+        tenantId,
+        userId,
+        ipAddress,
+        "deny_invalid_credentials",
+        riskLevel,
+        riskScore,
+        providerSignals
+      );
     }
 
     if (riskLevel == "high") {
@@ -467,7 +479,7 @@ class CDCService {
       if (toLower(providerSignals["arkose_result"].get!string) == "suspicious") score += 35;
     }
 
-    if ("transunion_score" in providerSignals && providerSignals["transunion_score"].type == Json.Type.integer) {
+    if ("transunion_score" in providerSignals && providerSignals["transunion_score"].type == Json.Type.int_) {
       if (providerSignals["transunion_score"].get!long > 700) score += 20;
     }
